Add first lesson materials
In the first lesson we will jump right in by first creating a
fairly generic app layout with HTML and CSS. After that
we'll get into JavaScript by building a "Camera" app that takes a
photo from the camera (or photo roll), displays that image and
overlays an effect on top of the image.
Then it's on to a "drawing app." Provided we have enough time
we'll include changing drawing colour, adding a settings a panel,
adjusting brush width, and also overlaying a "colouring page"
image.